+++ packages/webapp/src/single/usePreviewSize.ts
@@ -4,7 +4,7 @@ import useBodyDimensions from "../utils/useBodyDimensions";
 export const usePreviewSize = () => {
   const { height, width } = useBodyDimensions()
 
-  const previewWidth = useMemo(() => Math.max(width, height), [height, width])
+  const previewWidth = useMemo(() => Math.max(width, height) * (window.devicePixelRatio || 1), [height, width])
 
   return previewWidth
 }