+++ packages/common/src/fs/file-types.js
@@ -1,7 +1,7 @@
 const fileTypes = {
   image: 'ico,bmp,gif,png,jpg,jpeg,jpe,tif,tiff,thm,webp'.split(','),
   binImage: 'psd,xcf'.split(','),
-  rawImage: 'mrw,crw,cr2,cr3,dng,rw2,heic,heif'.split(','),
+  rawImage: 'arw,cr2,cr3,crw,dng,gpr,heic,heif,mrw,nef,nrw,orf,raf,rw2,sr2,srf,srw'.split(','),
   video: 'avi,wmv,mov,mpg,mpeg,mp4,m4v,3gp,mjpeg,mts,flv,swf,webm,mkv'.split(','),
   sound: 'wav,mp3,ogg,flac,m4a,aac,wma'.split(','),
   gps: 'gpx,kmz'.split(','),