+++ CHANGELOG.md
@@ -5,12 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.14.1] - 2023-11-20
 
 ### Fixed
 
-- cli: broken import-task after 'Keep config env for run import'
-
+- cli: Fix interactive import tasks
 
 ## [1.14.0] - 2023-11-18